--- components/shared/Dropdown/CarotIcon.tsx
@@ -1,17 +1,0 @@
-export function CarotIcon() {
-    return (
-      <svg
-        width="12"
-        height="10"
-        viewBox="0 0 12 10"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          className="carotIcon"
-          d="M6.86095 9.05668C6.47395 9.71206 5.52578 9.71206 5.13878 9.05668L0.731956 1.59372C0.338323 0.9271 0.818878 0.0852566 1.59304 0.0852566H10.4067C11.1809 0.0852566 11.6614 0.927101 11.2678 1.59372L6.86095 9.05668Z"
-        />
-      </svg>
-    )
-  }
-
